Move addDiagnosticFromNodeIfNotCached
