+++ src/TSTransformer/classes/DiagnosticService.ts
@@ -1,4 +1,5 @@
 import { hasErrors } from "Shared/util/hasErrors";
+import { TransformState } from "TSTransformer";
 import ts from "typescript";
 
 export class DiagnosticService {
@@ -10,6 +11,19 @@ export class DiagnosticService {
 
 	public static addDiagnostics(diagnostics: ReadonlyArray<ts.Diagnostic>) {
 		this.diagnostics.push(...diagnostics);
+	}
+
+	public static addDiagnosticFromNodeIfNotCached(
+		state: TransformState,
+		node: ts.Node,
+		diagnostic: ts.Diagnostic,
+		cache: Set<ts.Node | ts.Symbol>,
+	) {
+		const symbol = state.getOriginalSymbol(node);
+		if (!(symbol ? cache.has(symbol) : cache.has(node))) {
+			cache.add(symbol ?? node);
+			DiagnosticService.addDiagnostic(diagnostic);
+		}
 	}
 
 	public static flush() {
